Rename "URL Builder" to "URL Updater"
It's used to update an existing Build Service URL with the latest
versions of components, not to create/build a new Build Service URL
