+++ data/navigation.json
@@ -6,8 +6,8 @@
 			"current": false
 		},
 		{
-			"name": "URL Builder",
-			"href": "url-builder",
+			"name": "URL Updater",
+			"href": "url-updater",
 			"current": false
 		},
 		{
