+++ lib/build-service.js
@@ -12,9 +12,9 @@ module.exports = buildService;
 
 function buildService(options) {
 
-	const health = healthChecks(options);
-	options.healthCheck = health.checks();
-	options.goodToGoTest = health.gtg();
+	// const health = healthChecks(options);
+	// options.healthCheck = health.checks();
+	// options.goodToGoTest = health.gtg();
 	options.about = require('../about.json');
 	options.registry = new Registry(options);
