+++ lib/routes/url-updater.js
@@ -4,10 +4,10 @@ const express = require('express');
 const Raven = require('raven');
 const cacheControl = require('@financial-times/origami-service').middleware.cacheControl;
 const navigation = require('../../data/navigation.json');
-const parseModulesParameter = require('../url-builder/parse-modules-parameter');
-const evaluateModules = require('../url-builder/evaluate-modules');
-const parseBuildServiceUrl = require('../url-builder/parse-build-service-url');
-const updateUrlForResults = require('../url-builder/update-url-for-results');
+const parseModulesParameter = require('../url-updater/parse-modules-parameter');
+const evaluateModules = require('../url-updater/evaluate-modules');
+const parseBuildServiceUrl = require('../url-updater/parse-build-service-url');
+const updateUrlForResults = require('../url-updater/update-url-for-results');
 const UserError = require('../utils/usererror');
 
 function addNavigationToRequest() {
@@ -26,15 +26,15 @@ function addNavigationToRequest() {
 }
 
 module.exports = app => {
-    app.get('/url-builder/', cacheControl({ maxAge: '7d' }), addNavigationToRequest(), (request, response) => {
-        response.render('url-builder', {
+    app.get('/url-updater/', cacheControl({ maxAge: '7d' }), addNavigationToRequest(), (request, response) => {
+        response.render('url-updater', {
             title: 'Origami Build Service',
             layoutStyle: 'o-layout--landing',
             navigation: request.navigation
         });
     });
 
-    app.post('/url-builder/', cacheControl({ maxAge: '5m' }), addNavigationToRequest(), express.urlencoded({ extended: false }), express.text(), async (request, response) => {
+    app.post('/url-updater/', cacheControl({ maxAge: '5m' }), addNavigationToRequest(), express.urlencoded({ extended: false }), express.text(), async (request, response) => {
         try {
             const url = parseBuildServiceUrl(request.body['build-service-url']);
             const brand = url.searchParams.get('brand');
@@ -45,7 +45,7 @@ module.exports = app => {
             const updatedBuildServiceUrl = decodeURIComponent(
                 updateUrlForResults(url, results).toString()
             );
-            response.render('url-builder', {
+            response.render('url-updater', {
                 title: 'Origami Build Service',
                 layoutStyle: '',
                 navigation,
@@ -64,7 +64,7 @@ module.exports = app => {
                     + 'your Build Service URL against the latest component '
                     + 'releases.';
             }
-            return response.status(400).render('url-builder', {
+            return response.status(400).render('url-updater', {
                 title: 'Origami Build Service',
                 layoutStyle: '',
                 navigation: request.navigation,
