+++ lib/url-updater/update-url-for-results.js
