+++ views/url-updater.html
@@ -1,7 +1,7 @@
 <div class="o-layout__main" data-o-component="o-syntax-highlight">
 
 	<div class="o-layout-typography">
-		<h1>URL Builder</h1>
+		<h1>URL Updater</h1>
 	</div>
 
 	{{#if error}}
